refactor: use float64 for radom discount (#2821)

+++ services/skus/radom/radom.go
@@ -31,9 +31,9 @@ type Managed struct {
 }
 
 type Method struct {
-	Network            string `json:"network"`
-	Token              string `json:"token"`
-	DiscountPercentOff int64  `json:"discountPercentOff"`
+	Network            string  `json:"network"`
+	Token              string  `json:"token"`
+	DiscountPercentOff float64 `json:"discountPercentOff"`
 }
 
 type LineItem struct {
